backend/requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary
redis
python-dotenv
alembic
typer
requests
